+++ qgreenland/tasks/common/shapefile.py
@@ -5,29 +5,21 @@ import luigi
 
 from qgreenland.constants import TaskType
 from qgreenland.util.luigi import LayerTask
-from qgreenland.util.misc import temporary_path_dir
-from qgreenland.util.shapefile import (find_shapefile_in_dir,
-                                       reproject_shapefile,
-                                       subset_shapefile)
+from qgreenland.util.misc import find_single_file_by_ext, temporary_path_dir
+from qgreenland.util.shapefile import reproject_shapefile, subset_shapefile
 
-
-# TODO: Is there any task history? e.g. can we look at the final output target
-# and generate a list of tasks that were performed to generate it?
 
 class UnzipShapefile(LayerTask):
     task_type = TaskType.WIP
-    requires_task = luigi.Parameter()
-
-    def requires(self):
-        return self.requires_task
 
     def output(self):
         return luigi.LocalTarget(f'{self.outdir}/unzip/')
 
     def run(self):
-        with temporary_path_dir(self.output()) as temp_path:
-            zf = zipfile.ZipFile(self.input().path)
+        zf_path = find_single_file_by_ext(self.input().path, ext='.zip')
+        zf = zipfile.ZipFile(zf_path)
 
+        with temporary_path_dir(self.output()) as temp_path:
             for fn in zf.namelist():
                 zf.extract(fn, temp_path)
             zf.close()
@@ -35,10 +27,6 @@ class UnzipShapefile(LayerTask):
 
 class ReprojectShapefile(LayerTask):
     task_type = TaskType.WIP
-    requires_task = luigi.Parameter()
-
-    def requires(self):
-        return self.requires_task
 
     def output(self):
         # TODO: DRY this out. Place responsibility in LayerTask.
@@ -50,8 +38,7 @@ class ReprojectShapefile(LayerTask):
         return luigi.LocalTarget(out_dir)
 
     def run(self):
-        shapefile = find_shapefile_in_dir(self.input().path)
-
+        shapefile = find_single_file_by_ext(self.input().path, ext='.shp')
         gdf = reproject_shapefile(shapefile)
         with temporary_path_dir(self.output()) as temp_path:
             fn = os.path.join(temp_path, self.filename)
@@ -60,17 +47,12 @@ class ReprojectShapefile(LayerTask):
 
 class SubsetShapefile(LayerTask):
     task_type = TaskType.WIP
-    requires_task = luigi.Parameter()
-
-    def requires(self):
-        return self.requires_task
 
     def output(self):
         return luigi.LocalTarget(f'{self.outdir}/subset/')
 
     def run(self):
-        shapefile = find_shapefile_in_dir(self.input().path)
-
+        shapefile = find_single_file_by_ext(self.input().path, ext='.shp')
         gdf = subset_shapefile(shapefile)
         with temporary_path_dir(self.output()) as temp_path:
             fn = os.path.join(temp_path, self.filename)