+++ qgreenland/util/luigi.py
@@ -1,9 +1,10 @@
 import os
+import shutil
 
 import luigi
 
 from qgreenland.constants import CONFIG, TaskType
-from qgreenland.util.misc import get_layer_dir, get_layer_fn
+from qgreenland.util.misc import get_layer_dir, get_layer_fn, temporary_path_dir
 
 
 class LayerTask(luigi.Task):
@@ -13,8 +14,12 @@ class LayerTask(luigi.Task):
     pass a string instead of a whole config object as a parameter.
     """
 
+    requires_task = luigi.Parameter()
     layer_id = luigi.Parameter()
     task_type = None
+
+    def requires(self):
+        return self.requires_task
 
     @property
     def layer_cfg(self):
@@ -64,3 +69,12 @@ class LayerPipeline(luigi.Task):
 
     def output(self):
         return luigi.LocalTarget(get_layer_dir(self.cfg))
+
+    def run(self):
+        if os.path.isdir(self.input().path):
+            source_path = self.input().path
+        else:
+            source_path = os.path.dirname(self.input().path)
+
+        with temporary_path_dir(self.output()) as temp_path:
+            shutil.copytree(source_path, temp_path, dirs_exist_ok=True)